+++ src/main/java/com/mysql/MySqlSeatDao.java
@@ -16,6 +16,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManagerFactory;
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
@@ -107,7 +108,7 @@ public class MySqlSeatDao implements SeatDao {
     }
 
     @Override
-    public Seat getSeatById(int id) throws DaoException {
+    public Seat getSeatById(BigInteger id) throws DaoException {
         try {
             session = sessionFactory.openSession();
             session.beginTransaction();
